COrrected function parameters.
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@2358 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ libprocess/tip.h
@@ -44,8 +44,8 @@ typedef void (*GwyTipGuessFunc)(GwyDataField *data,
 				gdouble height,
 				gdouble radius,
 				gdouble *params,
-				gdouble *xres,
-				gdouble *yres);
+				gint *xres,
+				gint *yres);
 
 typedef struct _GwyTipModelPreset GwyTipModelPreset;
